move specific sre services into their own folder
